pyproject.toml
[project]
name = "p5_medical_db"
version = "0.1.0"
description = "Projet de formation N°5: Maintenez et documentez un système de stockage des données sécurisé et performant"
authors = [
    {name = "Abdulkadir GUVEN",email = "abguven@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo (>=4.15.1,<5.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
